# frontend/src/api/generated/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "description": "稳定错误码，代码里以此分支，不要解析文案",
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "success": {
            "const": false,
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "error",
          "code"
        ],
        "type": "object"
      },
...
      "StyleListResponse": {
        "properties": {
          "aspect_presets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array"
          },
          "categories": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "StyleEntry"
            },
            "type": "array"
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "success",
          "items"
        ],
        "type": "object"
      },
      "StyleResolveResponse": {
        "properties": {
          "match_kind": {
            "enum": [
              "id",
              "label",
              "alias",
              "substring"
            ],
            "type": "string"
          },
          "resolved": {
            "description": "false = 未命中。调用方**必须**继续按自由文本处理，不得凭空猜一个风格。",
            "type": "boolean"
          },
          "style_id": {
            "type": "string"
          },
          "success": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "resolved"
        ],
        "type": "object"
      },
      "StyleResponse": {
        "properties": {
          "item": {
            "$ref": "StyleEntry"
          },
          "success": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "本规格是**前端与后端的唯一契约**。前端客户端由 `scripts/generate_client.py` 从本文件生成，`scripts/check_generated_client.py` 在 CI 校验生成物是否过期。\n\n**手工改这个文件没有意义**：它是生成物，改了会被下一次生成覆盖。",
    "title": "漫剧锻造台 API",
    "version": "2026-10-07.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/styles": {
      "get": {
        "operationId": "listStyles",
        "parameters": [
          {
            "description": "按分类过滤：2d / 3d / real",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleListResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "风格库全量（后端单一事实源）",
        "tags": [
          "styles"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/styles/resolve": {
      "get": {
        "operationId": "resolveStyle",
        "parameters": [
          {
            "description": "项目 config.style 里存的原始字面值",
            "in": "query",
            "name": "text",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleResolveResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "由自由文本解析风格（存量项目向后兼容）",
        "tags": [
          "styles"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/styles/{style_id}": {
      "get": {
        "operationId": "getStyle",
        "parameters": [
          {
            "description": "稳定风格 id（改名不变）",
            "in": "path",
            "name": "style_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "风格不存在"
          }
        },
        "summary": "风格详情（stable style_id，改名不断历史）",
        "tags": [
          "styles"
        ],
        "x-contract-source": "manual"
      }
    },
...
  },
  "servers": [
    {
      "description": "同源部署",
      "url": "/"
    }
  ],
...
}